I made the deck card available to the player to take and to become unavailable after it is on player hand
